requirements.txt
asgiref==3.8.1
deepdiff==8.4.2
Django
django-cors-headers==4.7.0
djangorestframework==3.16.0
orderly-set==5.4.0
pillow==11.1.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.13.0
#openai
#pgvector
psycopg2-binary
drf-yasg==1.21.0
influxdb-client==1.36.0
django-health-check==3.17.0
celery==5.3.4
redis==5.0.1

